+++ src/components/common/buttons/bookmarkButton.tsx
@@ -1,8 +1,8 @@
 import Checkbox from '@mui/material/Checkbox'
 import BookmarkBorderIcon from '@mui/icons-material/BookmarkBorder'
 import BookmarkIcon from '@mui/icons-material/Bookmark'
-import createBookmark from 'services/bookmark/createBookmark'
-import deleteBookmark from 'services/bookmark/deleteBookmark'
+import createBookmark from 'services/snsPost/createBookmark'
+import deleteBookmark from 'services/snsPost/deleteBookmark'
 
 // TODO: 북마크 버튼 눌렀을 때 UI에 느리게 반영되는 것 개선하기
 const BookmarkButton = ({
